requirements.txt
python-telegram-bot[rate-limiter]~=21.4
aiohttp~=3.9
python-dotenv~=1.0

